requirements.txt
apturl==0.5.2
asgiref==3.4.1
attrs==20.3.0
Babel==2.9.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.6.20
cfgv==3.3.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.5.0
distro-info==1.0
Django==3.2.6
django-crispy-forms==1.12.0
django-extra-views==0.14.0
django-money==2.0.1
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.20
fasteners==0.14.1
filelock==3.4.2
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.4
httplib2==0.18.1
identify==2.4.4
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.9
nodeenv==1.6.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
pathspec==0.9.0
pexpect==4.8.0
Pillow==8.1.2
platformdirs==2.4.1
pre-commit==2.16.0
protobuf==3.12.4
py-moneyed==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycups==2.0.1
PyDrive==1.3.1
pyflakes==2.4.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.15.0
python-magic==0.4.20
pytz==2021.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.5.66
requests==2.25.1
rsa==4.0
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
systemd-python==234
texttable==1.6.3
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0
wadllib==1.3.5
websocket-client==0.57.0
whitenoise==5.3.0
xdg==5
xkit==0.0.0
zipp==1.0.0
